+++ RELEASES.md
@@ -4,6 +4,7 @@ Version 0.8.5 (2021-XX-XX)
  * Smart model-based Forward Collision Warning
  * Hyundai Elantra 2021 support thanks to CruiseBrantley!
  * Lexus UX Hybrid 2019 support thanks to brianhaugen2!
+ * Škoda Octavia 2015 & 2019 support thanks to jyoung8607!
 
 Version 0.8.4 (2021-05-17)
 ========================